server/requirements.txt
Django==1.11.6
djangorestframework==3.7.3
Pillow==5.0.0
mysqlclient==1.3.12
uwsgi==2.0.15
django-cors-headers==2.2.0
